Fix floating-point division

+++ app/src/main/java/io/github/formular_team/formular/math/CubicBezierCurve.java
@@ -59,7 +59,7 @@ public final class CubicBezierCurve implements Curve {
 
         for ( int d = 0; d <= divisions; d ++ ) {
 
-            points.push( this.getPoint( d / divisions ) );
+            points.push( this.getPoint( (float) d / divisions ) );
 
         }
 
@@ -77,7 +77,7 @@ public final class CubicBezierCurve implements Curve {
 
         for ( int d = 0; d <= divisions; d ++ ) {
 
-            points.push( this.getPoint( d / divisions ) );
+            points.push( this.getPoint( (float) d / divisions ) );
 
         }
