+++ app/src/main/java/io/github/formular_team/formular/math/Path.java
@@ -109,7 +109,7 @@ public class Path implements Curve{
 
         for ( int d = 0; d <= divisions; d ++ ) {
 
-            points.add(0, this.getPointAt( d / divisions ) );
+            points.add(0, this.getPointAt( (float) d / divisions ) );
 
         }
